+++ src/main/java/com/scaler/bookmyshow/controllers/BookingController.java
@@ -1,0 +1,15 @@
+package com.scaler.bookmyshow.controllers;
+
+import com.scaler.bookmyshow.dtos.CreateBookingRequestDto;
+import com.scaler.bookmyshow.dtos.CreateBookingResponseDto;
+import com.scaler.bookmyshow.services.BookingService;
+
+public class BookingController {
+    private BookingService bookingService;
+
+    public CreateBookingResponseDto createBooking(
+            CreateBookingRequestDto requestDto
+    ) {
+        return null;
+    }
+}