+++ src/main/java/com/scaler/bookmyshow/models/Booking.java
@@ -1,6 +1,8 @@
 package com.scaler.bookmyshow.models;
 
+import java.util.Date;
 import java.util.List;
+import java.util.zip.DataFormatException;
 
 public class Booking extends BaseModel {
     private User bookedBy;
@@ -9,6 +11,16 @@ public class Booking extends BaseModel {
     private double totalAmount;
     private List<Payment> payments;
     private BookingStatus bookingStatus;
+
+    private Date timeOfBooking;
+
+    public Date getTimeOfBooking() {
+        return timeOfBooking;
+    }
+
+    public void setTimeOfBooking(Date timeOfBooking) {
+        this.timeOfBooking = timeOfBooking;
+    }
 
     public User getBookedBy() {
         return bookedBy;
@@ -58,3 +70,6 @@ public class Booking extends BaseModel {
         this.bookingStatus = bookingStatus;
     }
 }
+
+// if there is a conflict of booking, allow the request
+// that is booking more seats