+++ src/main/java/com/scaler/bookmyshow/repositories/AuditoriumRepository.java
@@ -1,4 +1,6 @@
 package com.scaler.bookmyshow.repositories;
 
+import com.scaler.bookmyshow.models.Auditorium;
+
 public interface AuditoriumRepository {
 }