+++ src/main/java/com/scaler/bookmyshow/services/ShowSeatsConcurrencyManagerService.java
@@ -1,0 +1,9 @@
+package com.scaler.bookmyshow.services;
+
+import com.scaler.bookmyshow.models.ShowSeat;
+
+import java.util.List;
+
+public interface ShowSeatsConcurrencyManagerService {
+    boolean getLockForShowSeats(List<Long> showSeatIds);
+}